Move types into the package 'core'
