+++ v2/internal/reconcilers/error_classifier.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Azure/go-autorest/autorest/to"
 
 	"github.com/Azure/azure-service-operator/v2/internal/genericarmclient"
+	"github.com/Azure/azure-service-operator/v2/pkg/genruntime/core"
 )
 
 const (
@@ -24,11 +25,11 @@ func stringOrDefault(str string, def string) string {
 	return str
 }
 
-func ClassifyCloudError(err *genericarmclient.CloudError) (genericarmclient.CloudErrorDetails, error) {
+func ClassifyCloudError(err *genericarmclient.CloudError) (core.CloudErrorDetails, error) {
 	if err == nil || err.InnerError == nil {
 		// Default to retrying if we're asked to classify a nil error
-		result := genericarmclient.CloudErrorDetails{
-			Classification: genericarmclient.ErrorRetryable,
+		result := core.CloudErrorDetails{
+			Classification: core.ErrorRetryable,
 			Code:           UnknownErrorCode,
 			Message:        UnknownErrorMessage,
 		}
@@ -36,7 +37,7 @@ func ClassifyCloudError(err *genericarmclient.CloudError) (genericarmclient.Clou
 	}
 
 	classification := classifyInnerCloudError(err.InnerError)
-	result := genericarmclient.CloudErrorDetails{
+	result := core.CloudErrorDetails{
 		Classification: classification,
 		Code:           stringOrDefault(to.String(err.InnerError.Code), UnknownErrorCode),
 		Message:        stringOrDefault(to.String(err.InnerError.Message), UnknownErrorMessage),
@@ -44,7 +45,7 @@ func ClassifyCloudError(err *genericarmclient.CloudError) (genericarmclient.Clou
 	return result, nil
 }
 
-func classifyInnerCloudError(err *genericarmclient.ErrorResponse) genericarmclient.ErrorClassification {
+func classifyInnerCloudError(err *genericarmclient.ErrorResponse) core.ErrorClassification {
 	// See https://docs.microsoft.com/en-us/azure/azure-resource-manager/templates/common-deployment-errors
 	// for a breakdown of common deployment error codes. Note that the error codes documented there are
 	// the inner error codes we're parsing here.
@@ -52,7 +53,7 @@ func classifyInnerCloudError(err *genericarmclient.ErrorResponse) genericarmclie
 	code := to.String(err.Code)
 	if code == "" {
 		// If there's no code, assume we can retry on it
-		return genericarmclient.ErrorRetryable
+		return core.ErrorRetryable
 	}
 
 	switch code {
@@ -74,7 +75,7 @@ func classifyInnerCloudError(err *genericarmclient.ErrorResponse) genericarmclie
 		"ResourceNotFound",
 		"ResourceQuotaExceeded",
 		"SubscriptionNotRegistered":
-		return genericarmclient.ErrorRetryable
+		return core.ErrorRetryable
 	case "BadRequestFormat",
 		"Conflict",
 		// TODO: See https://github.com/Azure/azure-service-operator/issues/1997 for why this is commented out
@@ -99,10 +100,10 @@ func classifyInnerCloudError(err *genericarmclient.ErrorResponse) genericarmclie
 		"ReservedResourceName",
 		"SkuNotAvailable",
 		"SubscriptionNotFound":
-		return genericarmclient.ErrorFatal
+		return core.ErrorFatal
 	default:
 		// TODO: We could technically avoid listing the above Retryable errors since that's the default anyway
 		// If we don't know what the error is, default to retrying on it
-		return genericarmclient.ErrorRetryable
+		return core.ErrorRetryable
 	}
 }
