+++ v2/pkg/genruntime/core/cloud_error_details.go
@@ -3,7 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package genericarmclient
+package core
 
 type CloudErrorDetails struct {
 	Classification ErrorClassification
